implemented getting next 7 days forecast weather in activity

+++ app/src/main/java/dev/jaym21/mausam/data/remote/models/entities/FeelsLike.kt
@@ -13,5 +13,5 @@ data class FeelsLike(
     @Json(name = "eve")
     val eve: Double?,
     @Json(name = "morn")
-    val morn: Int?
+    val morn: Double?
 )
