+++ app/src/main/java/dev/jaym21/mausam/data/remote/models/entities/Temp.kt
@@ -17,5 +17,5 @@ data class Temp(
     @Json(name = "eve")
     val eve: Double?,
     @Json(name = "morn")
-    val morn: Int?
+    val morn: Double?
 )
